4주차

+++ todo/views.py
@@ -3,6 +3,7 @@ from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from django.views.decorators.csrf import csrf_exempt
 from todo.models import Todo
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -15,18 +16,23 @@ def index(request):
         }
         return render(request, "todo/index.html", context)
     else:
-        return HttpResponse('error', status=405)
+        return HttpResponse('Invalid request method', status=405)
 
 
+@login_required(login_url='/user/login/')
 @csrf_exempt
 def create(request):
     if request.method == "POST":
-        Todo.objects.create(content=request.POST['content'])
+        Todo.objects.create(
+            content=request.POST['content'],
+            user=request.user,
+            image=request.FILES.get("image"),
+        )
         return redirect('/todo/')
     elif request.method == "GET":
         return render(request, "todo/create.html")
     else:
-        return HttpResponse('error', status=405)
+        return HttpResponse('Invalid request method', status=405)
 
 
 def read(request, todo_id):
@@ -41,19 +47,25 @@ def read(request, todo_id):
 def delete(request, todo_id):
     if request.method == 'POST':
         todo = Todo.objects.get(id=todo_id)
-        todo.delete()
-        return redirect('/todo/')
+        if request.user == todo.user:
+            todo.delete()
+            return redirect('/todo/')
+        else:
+            return HttpResponse('not allowed to delete', status=403)
     else:
-        return HttpResponse('error', status=405)
+        return HttpResponse('Invalid request method', status=405)
 
 
 @csrf_exempt
 def update(request, todo_id):
     if request.method == "POST":
-        todo = Todo.objects.get(id=todo_id)
-        todo.content = request.POST['content']
-        todo.save()
-        return redirect(f'{todo_id}')
+        if request.user == todo.user:
+            todo = Todo.objects.get(id=todo_id)
+            todo.content = request.POST['content']
+            todo.save()
+            return redirect(f'{todo_id}')
+        else:
+            return HttpResponse('not allowed to delete', status=403)
     elif request.method == "GET":
         todo = Todo.objects.get(id=todo_id)
         context = {
@@ -61,4 +73,4 @@ def update(request, todo_id):
         }
         return render(request, "todo/update.html", context)
     else:
-        return HttpResponse('error', status=405)
+        return HttpResponse('Invalid request method', status=405)
